requirements.txt
# Core RAG
langchain==0.3.25
langchain-community==0.3.23
langchain-groq==0.3.2
langchain-huggingface==0.1.2

# Vector DB
chromadb==0.6.3

# Document loading
pypdf==5.4.0
unstructured==0.17.2

# Embeddings (local, free)
sentence-transformers==4.1.0

# Frontend
streamlit==1.45.0

# Guardrails / PII
presidio-analyzer==2.2.358
presidio-anonymizer==2.2.358
spacy==3.8.4

# Monitoring & Evaluation
langsmith==0.3.42
ragas==0.2.15

# Utilities
python-dotenv==1.1.0
pandas==2.2.3
pydantic==2.11.4
